refactor: centralize feature preprocessing
